added middleware, session, usercontroller, authentication functionality, search functionality

+++ routes.php
@@ -1,6 +1,7 @@
 <?php
 
 use App\Controllers\ListingController;
+use App\Controllers\UserController;
 
 // return [
 //     '/' => 'controllers/home.php',
@@ -15,17 +16,24 @@ $router->get('/listings', [ListingController::class, 'index']);
 
 $router->post('/listings', [ListingController::class, 'store']);
 
-$router->get('/listings/create', [ListingController::class, 'create']);
-$router->get('/listings/edit/{id}', [ListingController::class, 'edit']);
+$router->get('/listings/create', [ListingController::class, 'create'])->addRouteMiddleware(['auth']);
+
+$router->get('/listings/edit/{id}', [ListingController::class, 'edit'])->addRouteMiddleware(['auth']);
 
 $router->get('/listings/{id}', [ListingController::class, 'show']);
 
+$router->put('/listings/{id}', [ListingController::class, 'update'])->addRouteMiddleware(['auth']);
 
-$router->put('/listings/{id}', [ListingController::class, 'update']);
+$router->delete('/listings/{id}', [ListingController::class, 'delete'])->addRouteMiddleware(['auth']);
 
-$router->delete('/listings/{id}', [ListingController::class, 'delete']);
+$router->get('/auth/register', [UserController::class, 'create'])->addRouteMiddleware(['guest']);
 
+$router->get('/auth/login', [UserController::class, 'login'])->addRouteMiddleware(['guest']);
 
+$router->post('/auth/login', [UserController::class, 'authenticate']);
+
+$router->post('/auth/register', [UserController::class, 'store']);
+$router->post('auth/logout', [UserController::class, 'logout'])->addRouteMiddleware(['auth']);
 
 // $router->get('/', 'controllers/home.php');
